requirements.txt
groq
requests
python-dotenv
google-genai